Bug #896854 soc: sprd: iommu: fix iommu 32-bit kernel build issue
[root cause  ] compile error
[changes     ] fix iommu 32-bit kernel build issue
[side effects]no
[self test   ]:yes
[download normally]:yes
[power on/off normally]:yes
[do common repository/branch inspection]:yes
[is there dependence]:no
[confirm dependent commit]:no
[board]:common
[test case]:compile pass
[reviewers   ] Wenan Hu
[change_type ] feature_bugfix
[tag_product ] common

fix iommu 32-bit kernel build issue.

+++ drivers/soc/sprd/iommu/drv/hal/sprd_iommuex_hal_register.c
@@ -146,7 +146,7 @@ void mmu_ex_first_ppn(ulong ctrl_base_addr, u32 iommu_id, ulong pp_addr)
 		reg_write_dword(reg_addr, pp_addr & 0xfffff000);
 		/*higher bits of pagetable base addr*/
 		reg_addr += 0x4;
-		reg_write_dword(reg_addr, (pp_addr >> 32) & 0xff);
+		reg_write_dword(reg_addr, ((u64)pp_addr >> 32) & 0xff);
 	} else {
 		reg_addr += FIRST_PPN_OFFSET;
 		reg_write_dword(reg_addr,
